Split out warning tests as they are order dependent
